refactor(webview): modularise TEC-1G entrypoint into feature modules (#228)
- Add entry-types, keypad-layout (declarative tables), tec1g-audio, tec1g-keypad, tec1g-platform-update
- Add tec1g-project-status-ui, tec1g-tab-memory, tec1g-memory-views; slim index.ts as composition root
- Add webview test for keypad layout tables

Fixes #228

Made-with: Cursor
